Rework FileUpload module on CRUD repository
