+++ packages/server-scala/src/main/scala/modules/upload/actors/FileActor.scala
@@ -4,9 +4,10 @@ import akka.actor.{Actor, ActorLogging, ActorRef}
 import akka.pattern._
 import com.google.inject.Inject
 import common.ActorNamed
+import common.DatabaseExecutor._
 import modules.upload.actors.FileActor.SaveFileMetadata
 import modules.upload.models.FileMetadata
-import modules.upload.repositories.FileMetadataRepo
+import modules.upload.repositories.FileMetadataRepository
 
 import scala.concurrent.ExecutionContext
 
@@ -17,12 +18,12 @@ object FileActor extends ActorNamed {
   final val name = "FileActor"
 }
 
-class FileActor @Inject()(fileRepo: FileMetadataRepo)
+class FileActor @Inject()(fileMetadataRepository: FileMetadataRepository)
                          (implicit val executionContext: ExecutionContext) extends Actor with ActorLogging {
 
   override def receive: Receive = {
     case saveFileMetadata: SaveFileMetadata =>
       log.info(s"Received a message: [ $saveFileMetadata ]")
-      fileRepo.create(saveFileMetadata.file).pipeTo(saveFileMetadata.sender)
+      fileMetadataRepository.save(saveFileMetadata.file).run.pipeTo(saveFileMetadata.sender)
   }
 }
