--- packages/server-scala/src/main/scala/modules/upload/guice/modules/ResolverModule.scala
@@ -1,11 +1,0 @@
-package modules.upload.guice.modules
-
-import modules.upload.graphql.resolvers.{FileUploadResolver, FileUploadResolverImpl}
-import net.codingwell.scalaguice.ScalaModule
-
-class ResolverModule extends ScalaModule {
-
-  override def configure() {
-    bind[FileUploadResolver].to[FileUploadResolverImpl]
-  }
-}
