+++ packages/server-scala/src/main/scala/modules/upload/models/FileMetadata.scala
@@ -1,32 +1,14 @@
 package modules.upload.models
-import slick.lifted.Tag
-import slick.jdbc.SQLiteProfile.api.{Table => SlickTable, _}
+
+import com.byteslounge.slickrepo.meta.Entity
 import spray.json.{DefaultJsonProtocol, JsonFormat}
 
 case class FileMetadata(id: Option[Int] = None,
                         name: String,
                         contentType: String,
                         size: Long,
-                        path: String)
-
-object FileMetadata extends ((Option[Int], String, String, Long, String) => FileMetadata) {
-
-  val name = "FILES"
-
-  class Table(tag: Tag) extends SlickTable[FileMetadata](tag, name) {
-
-    def id = column[Int]("ID", O.PrimaryKey, O.AutoInc)
-
-    def name = column[String]("NAME")
-
-    def contentType = column[String]("CONTENT_TYPE")
-
-    def size = column[Long]("SIZE")
-
-    def path = column[String]("PATH")
-
-    override def * = (id.?, name, contentType, size, path).mapTo[FileMetadata]
-  }
+                        path: String) extends Entity[FileMetadata, Int] {
+  override def withId(id: Int): FileMetadata = this.copy(id = Some(id))
 }
 
 object FileMetadataJsonProtocol extends DefaultJsonProtocol {
